All The requirements are been developed

+++ Task_API/Controllers/UserController.cs
@@ -31,50 +31,63 @@ namespace Task_API.Controllers
         public async Task<ActionResult<MUser>> AddUser(MUser muser)
         {
             var userRoles = User.Claims.Where(c => c.Type == ClaimTypes.Role).Select(c => c.Value).ToList();
-            try
+            string loggedinUser = HttpContext.User.FindFirstValue("UserName"); // code to get username who is loggedin 
+            var userIsValidOrNo = await _userRepository.UserIsActiveOrNot(loggedinUser);
+            if (userIsValidOrNo == "Active")
             {
-                var existingUser = await _userRepository.GetUserByName(muser.UUserName);
-                if (existingUser != null)
+                try
                 {
-                    return Conflict("Username already exixts...");
+                    var existingUser = await _userRepository.GetUserByName(muser.UUserName);
+                    if (existingUser != null)
+                    {
+                        return Conflict("Username already exixts...");
+                    }
+                    var addingUser = await _userRepository.AddUserAccount(muser);
+                    await _taskDataBaseContext.SaveChangesAsync();
+                    return CreatedAtAction(nameof(AddUser), addingUser);
                 }
-                var addingUser = await _userRepository.AddUserAccount(muser);
-                await _taskDataBaseContext.SaveChangesAsync();
-                return CreatedAtAction(nameof(AddUser), addingUser); 
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex.Message);
+                    return StatusCode(500, "Internal Server Error...");
+                }
             }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex.Message);
-                return StatusCode(500, "Internal Server Error...");
-            }
+            return StatusCode(404, "Your status is Inactive please contact Adminstratio...");
+
         }
 
 
-       
+
         [HttpGet]
         [Route("GetValue")]
         [Authorize]
         public async Task<ActionResult> getvalue(string username)
         {
             var userRoles = User.Claims.Where(c => c.Type == ClaimTypes.Role).Select(c => c.Value).ToList();
+            string loggedinUser = HttpContext.User.FindFirstValue("UserName"); // code to get username who is loggedin 
+            var userIsValidOrNo = await _userRepository.UserIsActiveOrNot(loggedinUser);
+            if (userIsValidOrNo == "Active")
+            {
+                try
+                {
+                    var use = await _userRepository.GetUserByName(username);
+                    if (use != null)
+                    {
+                        return StatusCode(200, use);
+                    }
+                    else
+                    {
+                        return StatusCode(404, "There is no user with this Username...");
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex.Message);
+                    return StatusCode(500, "Internal Server Error...");
+                }
+            }
+            return StatusCode(404, "Your status is Inactive please contact Adminstratio...");
 
-            try
-            {
-                var use = await _userRepository.GetUserByName(username);
-                if (use != null)
-                {
-                    return StatusCode(200, use);
-                }
-                else
-                {
-                    return StatusCode(404, "There is no user with this Username...");
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex.Message);
-                return StatusCode(500, "Internal Server Error...");
-            }
         }
 
 
@@ -82,16 +95,23 @@ namespace Task_API.Controllers
         [Route("UpdateUser")]
         public async Task<ActionResult> updateuser(TUser user)
         {
-            try
+            string loggedinUser = HttpContext.User.FindFirstValue("UserName"); // code to get username who is loggedin 
+            var userIsValidOrNo = await _userRepository.UserIsActiveOrNot(loggedinUser);
+            if (userIsValidOrNo == "Active")
             {
-                var updatedUser = await _userRepository.UpdateUserAccountByID(user);
-                return StatusCode(200, updatedUser);
+                try
+                {
+                    var updatedUser = await _userRepository.UpdateUserAccountByID(user);
+                    return StatusCode(200, updatedUser);
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex.Message);
+                    return StatusCode(500, "Internal Server Error...");
+                }
             }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex.Message);
-                return StatusCode(500, "Internal Server Error...");
-            }
+            return StatusCode(404, "Your status is Inactive please contact Adminstratio...");
+
         }
 
 
@@ -99,16 +119,23 @@ namespace Task_API.Controllers
         [Route("DeleteUser")]
         public async Task<ActionResult> delUser(int id)
         {
-            try
+            string loggedinUser = HttpContext.User.FindFirstValue("UserName"); // code to get username who is loggedin 
+            var userIsValidOrNo = await _userRepository.UserIsActiveOrNot(loggedinUser);
+            if (userIsValidOrNo == "Active")
             {
-                var dellUSer = await _userRepository.DeleteUser(id);
-                return StatusCode(200, "User Deleted Succesfully...");
+                try
+                {
+                    var dellUSer = await _userRepository.DeleteUser(id);
+                    return StatusCode(200, "User Deleted Succesfully...");
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex.Message);
+                    return StatusCode(500, "Internal Server Error...");
+                }
             }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex.Message);
-                return StatusCode(500, "Internal Server Error...");
-            }
+            return StatusCode(404, "Your status is Inactive please contact Adminstratio...");
+
         }
